requirements.txt
Flask==2.0.2
gitpython==3.1.24
schedule==1.1.0